Move the package model and mocked logic functions from views
They are still mocked and unfinished but I would like to use them
in other views as well so they cannot be local anymore.
